Add files via upload
Gender equality is today considered in many areas in society, including gender balance in movies. To be able to assess this, one could study different aspects of a movie that could indicate what gender the lead actor has. The aim of this project was to evaluate and compare different classifiers, trained to predict the gender of the two main actors in Hollywood movies based on features, such as the amount of dialogue the actors have or the release year. We Implemented different statistical machine learning models. Especially optimized the parameters and evaluated, by doing so it showed how much better a model can perform by just changing a few parameters both mathematically and computational. Also understood that data selection is equally important as implementing a good model.
The Algorithms implemented are Decision Tree and Random Forest Method.